% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{abr}
\alias{abr}
\title{Single ionisaiton mode of abr1 FIE-MS data (raw)}
\format{List object containing two elements
	\itemize{
		\item neg negative ionisation mode data frame (2000 variables)
		\item cls vector of class labels (N = 6)
	}}
\usage{
data(abr)
}
\description{
Single ionisaiton mode of abr1 FIE-MS data (raw)
}
\author{
Tom Wilson
}
\keyword{datasets}

