% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets}
\alias{assets}
\alias{format.assets}
\alias{print.assets}
\alias{[.assets}
\alias{as.double.assets}
\alias{names.assets}
\alias{as.data.frame.assets}
\title{Assets}
\usage{
assets(x)

\method{format}{assets}(x, ...)

\method{print}{assets}(x, ...)

\method{[}{assets}(x, i, ...)

\method{as.double}{assets}(x, ...)

\method{names}{assets}(x)

\method{as.data.frame}{assets}(x)
}
\arguments{
\item{x}{object of class 'assets'.}
}
\value{
An asset data frame with the name of the asset and all provided
additional information.
}
\description{
Each item can provide multiple assets. An asset specifies the type of data
(e.g., historical, recent) as well as the temporal resolution of the data
and the data period (if suitable). The \code{ms_assets} class is used to handle
these assets within the package.
}
\author{
Reto
}
