% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessapi.R
\name{autoconvert_datetime}
\alias{autoconvert_datetime}
\title{Converting Datetime Columns}
\usage{
autoconvert_datetime(x)
}
\arguments{
\item{x}{data frame.}
}
\value{
Data frame with POSIXct variables if any variable
containing datetime information was detected.
}
\description{
Takes a data frame as input and tries to identify columns
containing datetime information.
}
\author{
Reto
}
