% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PeakFilter.R
\docType{methods}
\name{BasicFilter}
\alias{BasicFilter}
\title{Basic peak filtering}
\usage{
\S4method{BasicFilter}{MSdata}(msdata, method = "none")
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object to be filtered}

\item{method}{Method of filtering one of:\cr
\code{"none"} - no filtering applied\cr
\code{"iqr"} - interquantile range (IQR)\cr
\code{"sd"} - standard deviation (SD)\cr
\code{"mad"} - median absolute deviation (MAD)\cr
\code{"rsd"} - relative standard deviation (RSD = SD/mean)\cr
\code{"nprsd"} - non-parametric relative standard deviation (MAD/median)\cr
\code{"mean"} - mean intensity value\cr
\code{"median"} - median intensity value
none (less than 2000 features)
the final variable should be less than 5000 for effective computing}
}
\value{
\code{\link{MSdata-class}} object without filtered peaks
}
\description{
The purpose of the data filtering is to identify and remove variables that are unlikely
to be of use when modeling the data. No phenotype information are used in the filtering process,
so the result can be used with any downstream analysis.
This step is strongly recommended for untargeted metabolomics datasets
(i.e. spectral binning data, peak lists) with large number of variables,
many of them are from baseline noises. Filtering can usually improve the results.\cr
}
\details{
Non-informative variables can be characterized in two groups: \itemize{
\item variables of very small values - can be detected using mean or median;
\item variables that are near-constant throughout the experiment conditions -
can be detected using different variance measures.
}

The following empirical rules are applied during data filtering:\itemize{
\item less than 250 variables: 5\% will be filtered;\cr
\item between 250 - 500 variables: 10\% will be filtered;\cr
\item between 500 - 1000 variables: 25\% will be filtered;\cr
\item over 1000 variables: 40\% will be filtered.
}

Please note, that \code{"none"} option is only for less than 2000 features.
Over that, if you choose \code{"none"}, the  IQR filter will still be applied.

The maximum allowed number of variables is 5000.
If over 5000 variables were left after filtering, only the top 5000 will be used in the
subsequent analysis.
}

