% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Norm_Scaling.R
\docType{methods}
\name{DataScaling}
\alias{DataScaling}
\title{Data scaling}
\usage{
\S4method{DataScaling}{MSdata}(msdata, method = "pareto")
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{The method of scaling, one of:\cr
\code{"auto"} - Autoscaling, mean-centring and dividing by the standard deviation of each variable;\cr
\code{"pareto"} - Pareto scaling, mean-centring and dividing by the square root of standard deviation of each variable; \cr
\code{"range"} - Range scaling, mean-centring and dividing by the range of each variable}
}
\value{
\code{\link{MSdata-class}} object with normalised intensity matrix
}
\description{
This function realizes two steps: mean-centring and scaling.
Mean-centring means that - for each feature (peak/compound) all samples intensities are considered
as differences from the mean intensity of this feature.
Scaling means that these differences are standardized for all the features by dividing by
standard deviation, or range, or another measure of variance.
Therefore, after scaling all the features have mean intensity 0 and corresponding variance measure 1.
}

