% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Norm_Transform.R
\docType{methods}
\name{DataTransform,MSdata-method}
\alias{DataTransform,MSdata-method}
\title{Data transformation}
\usage{
\S4method{DataTransform}{MSdata}(msdata, method = "glog10")
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{The method of transformation, one of:\cr
\code{"log10"} - log10-transformation; \cr
\code{"log2"} - log2-transformation;\cr
\code{"glog10"} - generalised log10, tolerant to zeros (zeros are replaced with 1/10 of the minimal value);\cr
\code{"glog2"} - generalised log2, tolerant to zeros (zeros are replaced with 1/10 of the minimal value); \cr
\code{"cuberoot"} - cube root transformation \cr}
}
\value{
\code{\link{MSdata-class}} object with transformed intensity matrix
}
\description{
Logarithmical or cube root data transformation.
}

