% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EvalMissVal.R
\docType{methods}
\name{EvalMissVal}
\alias{EvalMissVal}
\title{Evaluate Missing Values}
\usage{
\S4method{EvalMissVal}{MSdata}(msdata, method = "Min", percent = 0.5)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}}}

\item{method}{Method of evaluation, one of:
\code{"Min"},\cr
\code{"FeatureMin"}, \code{"FeatureMean"}, \code{"FeatureMedian"}\cr
\code{"GausSim"}\cr
\code{"knn"}\cr
\code{"bpca"}, \code{"ppca"} or \code{"svdImpute"}}

\item{percent}{For \code{method = "Min"}. Quotient of minimal value used to replace missing values.}
}
\description{
Fill NA positions in data set. See more: \code{\link[pcaMethods]{pca}}
}

