% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{FIEHRMSraw}
\alias{FIEHRMSraw}
\title{Parsed FIE-HRMS mzXML}
\format{List object containing 2 elements; \code{positive} and \code{negative}, each with 24 sub-elements. The sub-elements
of each ionisation mode are the raw parsed output of \code{mzXML} files using \code{FIEtools::readFIE}}
\usage{
data(FIEHRMSraw)
}
\description{
Parsed FIE-HRMS mzXML
}
\author{
Tom Wilson
}
\keyword{datasets}

