% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_request}
\alias{get_request}
\title{Get Request Helper Function}
\usage{
get_request(url, query = NULL, ...)
}
\arguments{
\item{url}{character, URL to send the request to. Can contain
get parameters in the URL.}

\item{query}{\code{NULL} (default) or a list with get parameters to be sent
with the request. Warning: If the URL contains get requests and
\code{query} is not \code{NULL}, the parameters in the URL (\code{url}) will
be overwritten!}
}
\value{
List with the decoded information.
}
\description{
Auxilary function to send GET requests with additional error handling.
}
\details{
We are expecting a JSON response from the API, thus
we expect that the request content is decoded to a list by the
httr package. If we do not get a proper HTTP status code, or
the extracted content is not a list, this function will throw
an error.
}
\author{
Reto
}
