% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{ms_stations}
\alias{ms_stations}
\title{Available Automated Weather Stations}
\usage{
ms_stations(url = NULL)
}
\arguments{
\item{url}{\code{NULL} or an URL TODO: If we need to retrieve
data from other collections we may better go for a 'collection'
argument, for now only tested \code{ch.meteoschweiz.ogd-smn} which is
specified internally.}
}
\value{
A simple feature data frame with station name, id, and its
geographical location. Unfortunately, we do not get altitude information
or information about the period a specific station provides information.
TODO: Currently ignoring 'assets' which contain the CSV file names
containing information about temporal resolution and type of data.
We will need to extract that.
See \url{https://opendatadocs.meteoswiss.ch/general/download}.
}
\description{
Retrieving all available automatic weather stations.
}
\author{
Reto
}
