% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSdata_class_basic_methods.R
\docType{class}
\name{MSdata-class}
\alias{MSdata}
\alias{MSdata-class}
\alias{intMatrix,MSdata-method}
\alias{intMatrix<-,MSdata-method}
\alias{peakData,MSdata-method}
\alias{peakData<-,MSdata-method}
\alias{sampleData,MSdata-method}
\alias{sampleData<-,MSdata-method}
\title{MSdata class}
\usage{
\S4method{peakData}{MSdata}(msdata)

\S4method{peakData}{MSdata}(msdata) <- value

\S4method{sampleData}{MSdata}(msdata)

\S4method{sampleData}{MSdata}(msdata) <- value

\S4method{intMatrix}{MSdata}(msdata)

\S4method{intMatrix}{MSdata}(msdata) <- value
}
\description{
MSdata-S4 class description, with it's accessors and replacement methods.
}
\section{Slots}{

\describe{
\item{\code{intMatrix}}{The matrix of peak intensities / compound concentrations.}

\item{\code{peakData}}{Peak metadata.}

\item{\code{sampleData}}{Sample metadata.}

\item{\code{processLog}}{Processing log.}
}}

