% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSdata_to_MA.R
\docType{methods}
\name{MSdata_to_MA}
\alias{MSdata_to_MA}
\title{Convert MSdata object to MetaboAnalyst object}
\usage{
\S4method{MSdata_to_MA}{MSdata}(msdata, designType = "regular",
  facA = names(sampleData(msdata))[1], facB = NULL)
}
\arguments{
\item{msdata}{An object of \code{\link[MSdata]{MSdata}} class.}

\item{designType}{\code{"time"} if the data is time-series data; \code{"regular"} otherwise. 
In the first case one of the arguments \code{facA} or \code{facB} have to be equal \code{"Time"}}

\item{facA}{Grouping factor, one of the names of sample data columns in \code{msdata@sampleData}}

\item{facB}{Optional second factor.}
}
\description{
Create an object for storing data for processing in MetaboAnalysis (\code{MApckg}).
}

