% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingValues.R
\docType{methods}
\name{msFillNA}
\alias{msFillNA}
\title{Evaluate Missing Values}
\usage{
\S4method{msFillNA}{MSdata}(msdata, method = "min", percent = 0.5)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}}}

\item{method}{Method of evaluation, one of:
\itemize{
\item \code{"min"}, - certain \code{percent} of minimal intensity throughout all dataset;\cr
\item \code{"featureMin"}, \code{"featureMean"}, \code{"featureMedian"} - NAs in feature are replaced by corresponding statistics of this feature intensity values;\cr
\item \code{"GausSim"} - random generation of normally distributed values;\cr
\item \code{"knn"} - nearest neighbour averaging. \code{\link[impute]{impute.knn}} function is used;\cr
\item \code{"bpca"}, \code{"ppca"}, \code{"svdImpute"} - imputation based on PCA analysis. \code{\link[pcaMethods]{pca}} function is used.
}}

\item{percent}{For \code{method = "min"}. Quotient of minimal value used to replace missing values.}
}
\description{
Fill NA positions in data set. See more: \code{\link[pcaMethods]{pca}}
}
\seealso{
\code{\link{msNoiseGen}}
}

