% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm.R
\docType{methods}
\name{msNorm}
\alias{msNorm}
\title{MSdata normalisation}
\usage{
\S4method{msNorm}{MSdata}(msdata, method, ref.cmpd = NULL)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{Method of normalisation. In each sample data are normalised by one of:\cr
\code{"sum"} -  sum concentration of all compounds\cr
\code{"median"} - median concentration of all compounds\cr
\code{"refCompound"} - concentration of reference compound\cr
\code{"robustPeaks"} - normalization by robust peaks (see details in Meissner, Steinhauser, Dittmann 2015)}

\item{ref.cmpd}{For \code{"refCompound"} method, the name or the order number of reference compound.}
}
\value{
\code{\link{MSdata-class}} object with normalised intensity matrix
}
\description{
Sample-wise normalisation methods.
}
\seealso{
\code{\link{msNormStandard}}, \code{\link{msNormBiomass}}, \code{\link{msScaling}}, \code{\link{msTransform}}
}

