% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormBiomass.R
\docType{methods}
\name{msNormBiomass}
\alias{msNormBiomass}
\title{Normalisation by biomass}
\usage{
\S4method{msNormBiomass}{MSdata}(msdata, biomass.list)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{biomass.list}{One of: 
\enumerate{
\item a path to the file containing the simple table with two columns: sample ID and biomass/volum/etc.
\item a name of the corresponding column in sample data table \code{sampleData(msdata)} (if biomass were uploaded there)
}}
}
\value{
\code{\link{MSdata-class}} object with normalised intensity matrix
}
\description{
Normalisation of intensities by the list of sample masses/volumes/etc.
}

