% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.R
\docType{methods}
\name{msOutlierMulti,MSdata-method}
\alias{msOutlierMulti,MSdata-method}
\title{Detect Multivariate outliers}
\usage{
\S4method{msOutlierMulti}{MSdata}(msdata, method = "pcdist", ...)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{Method of outlier detection, one of:\cr
\code{"mahdist"} - based on the Mahalanodis distance (\code{\link[rrcovHD]{OutlierMahdist}} is used)\cr
\code{"pcdist"} - algorithm proposed by Shieh and Hung (2009) (\code{\link[rrcovHD]{OutlierPCDist}} is used)\cr
\code{"pcout"} - algorithm proposed by Filzmoser, Maronna and Werner (2008) (\code{\link[rrcovHD]{OutlierPCOut}} is used)\cr
\code{"sign1"} - using spatial sings, computation is based on the Mahalanodis distance (\code{\link[rrcovHD]{OutlierSign1}} is used)\cr
\code{"sign2"} - using spatial sings, computation is based on principal components (\code{\link[rrcovHD]{OutlierSign2}} is used)}

\item{...}{- other arguments passed to the function of the respective method}
}
\description{
The function detects multivariate outliers among samples.
}
\seealso{
\code{\link{msOutlierUni}}, \code{\link{msOutlierMulti}}
}

