% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.R
\docType{methods}
\name{msOutlierPeakCV,MSdata-method}
\alias{msOutlierPeakCV,MSdata-method}
\title{Detect Outlying Peaks}
\usage{
\S4method{msOutlierPeakCV}{MSdata}(msdata)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{Method of outlier detection, one of:\cr
\code{"tukey"} - Tukey's biweight.
\code{"box"} - boxplot statistic; values more than 
(75% quantile + 1.5*IQR) or less than (25% quantile - 1.5*IQR) 
are considered as outliers. If the number of values is less than 6,
only the most outlying one will be removed.
\code{"both"} - only values detected by both algorithms are marked as outliers.}
}
\description{
The function detects reliably quantified peaks using their groupwise 
coefficient of variation (CV). For this, the mean average, the Tukey's biweight,
and the median over all groupwise CV values are estimated and outliers 
are removed by iterative boxplot statistics. 
Peaks revealing outlying values in the majority of cases (2 out of 3) are 
removed from further analysis.
}
\seealso{
\code{\link{msOutlierUni}}, \code{\link{msOutlierPeakCV}}
}

