% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier.R
\docType{methods}
\name{msOutlierUni,MSdata-method}
\alias{msOutlierUni,MSdata-method}
\title{Detect Univariate Outliers}
\usage{
\S4method{msOutlierUni}{MSdata}(msdata, method = "both")
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{Method of outlier detection, one of:\cr
\code{"tukey"} - Tukey's biweight.
\code{"box"} - boxplot statistic; values higher than 
(75% quantile + 1.5*IQR) or lower than (25% quantile - 1.5*IQR) 
are considered as outliers. If the number of values is less than 6,
only the most outlying one will be removed.
\code{"both"} - only values detected by both algorithms are marked as outliers.}
}
\description{
The function detects univariate outliers among peak intensity values
within each replication group. Outlying values are replaced with NAs.
}
\seealso{
\code{\link{msOutlierMulti}}, \code{\link{msOutlierPeakCV}}
}

