% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSoutput.R
\docType{methods}
\name{MSoutput}
\alias{MSoutput}
\title{Output MSdata in files}
\usage{
\S4method{MSoutput}{MSdata}(msdata, dir = "", file = "MSdata")
}
\arguments{
\item{msdata}{An object of \code{\link[MSdata]{MSdata}} class.}

\item{dir}{The output directory path.}

\item{file}{File name prefix.}
}
\description{
Writes MSdata to four files, containing intensity matrix, sample metadata,
peak metadata and processing log file.
}

