% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSupload.R
\docType{methods}
\name{MSupload}
\alias{MSupload}
\alias{MSupload,character-method}
\alias{MSupload,list-method}
\title{Upload MSdata}
\usage{
MSupload(object, ...)

\S4method{MSupload}{character}(object, orientation = "SamplesInCol",
  zeros.as.NA = TRUE, sampleDataLines = 1, peakDataLines = 1,
  sampleNames = TRUE, peakNames = TRUE)

\S4method{MSupload}{list}(object = list(intFile = "", sampleFile = "",
  peakFile = ""), orientation = "SamplesInCol", zeros.as.NA = TRUE)
}
\arguments{
\item{object}{One of:
\enumerate{
\item a character vector of length 1 - just a file path name of one .csv or .txt data frame.
\item a list of paths to three files: matrix of intensities, sample metadata and peak metadata.
}}

\item{orientation}{Orientation of table, one of \code{"SamplesInCol"} or \code{"SamplesInRow"}}

\item{zeros.as.NA}{If \code{TRUE} then zero intensities are treated as NAs.}

\item{sampleDataLines}{The number of lines containing data about samples.}

\item{peakDataLines}{The number of lines containing data about peaks/compounds.}

\item{sampleNames}{If \code{TRUE}, sample names are taken from the first column/row of table. \cr
If \code{FALSE}, standard names like \code{"sample1"} are created.}

\item{peakNames}{If \code{TRUE}, sample names are taken from the first row/column of table. \cr
If \code{FALSE}, standard names like \code{"peak1"} are created.}
}
\value{
\code{\link{MSdata-class}} object
}
\description{
Create a \code{\link{MSdata-class}} object from external data tables.
Automatically adds a \code{ReplicationGroup} column in \code{sampleData} table according to group factors combinations.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{list}: 
}}

