% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Norm.R
\docType{methods}
\name{Norm,MSdata-method}
\alias{Norm,MSdata-method}
\title{MSdata normalisation}
\usage{
\S4method{Norm}{MSdata}(msdata, method, biomass.list = NULL,
  ref.cmpd = NULL)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{method}{Method of normalisation. In each sample data are normalised by one of:\cr
\code{"Sum"} -  sum concentration of all compounds\cr
\code{"Median"} - median concentration of all compounds\cr
\code{"RefCompound"} - concentration of reference compound\cr
\code{"Biomass"} - sample biomass/volume/etc. \cr}

\item{biomass.list}{For \code{"Biomass"} method, the link to the file containing the simple table with two columns: sampleID and biomasses/volumes/etc.}

\item{ref.cmpd}{For \code{"RefCompound"} method, the name or number of reference compound.}
}
\value{
\code{\link{MSdata-class}} object with normalised intensity matrix
}
\description{
Sample-wise normalisations.
}

