% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PeakFilter.R
\docType{methods}
\name{PeakFilter}
\alias{PeakFilter}
\title{Peak filtering}
\usage{
\S4method{PeakFilter}{MSdata}(msdata, blanks = NULL, above.blank = 2,
  min.int = 1000, min.nonNAnum.repgroup = 3, min.nonNApercent = 0.4)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object to be filtered}

\item{blanks}{The vector of blank samples. Either vector of sample numbers or sample names.
(check names by command: \code{sampleData(msdata)}).
These sample will be removed from dataset after filtering.}

\item{above.blank}{Filter peaks with intensities close to blanks.
\code{mean - above.blank*SE} for certain peak's intensities in replicate group have to be more than
\code{mean + above.blank*SE} for this peak's intensities in blank group.
(where \code{SE} - standard error of the mean.)\cr
Note: if \code{above.blank = 0} then just mean values are compared.}

\item{min.int}{Filter peaks by total intensity. Mean peak intensity in replicate group
have to be higher than \code{min.int} value.}

\item{min.nonNAnum.repgroup}{Filter peaks with too many missing values. \code{min.number.repgroup}
is minimal number of non-NA values in replicate group.}

\item{min.nonNApercent}{Filter peaks with too many missing values. \code{min.allgroups}
is minimal allowed quotient of non-NA values for each peak.}
}
\value{
\code{\link{MSdata-class}} object without filtered peaks and blank samples
}
\description{
Remove peaks fitting some criteria from data set. Criteria are stated by function arguments.
If an argument is equal \code{NULL}, corresponding filtering criterion is not applied.\cr
Usually each peak in each replicate group considered separately in terms of each criterion.
After that this peak/replicate group pair can be marked as "filterable" according to this criterion.
If for at least one selecting criterion all replicate groups are marked as "filterable",
this peak is removed from data set.
}

