% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{pw1HRMS}
\alias{pw1HRMS}
\title{Binary High Resolution FIE-MS data set (processed)}
\format{List object containing three elements
	\itemize{
		\item pos positive ionisation mode data frame (3079 variables)
		\item neg negative ionisation mode data frame (1452 variables)
		\item cls vector of class labels (N = 2)
	}}
\usage{
data(pw1HRMS)
}
\description{
Binary High Resolution FIE-MS data set (processed)
}
\author{
Tom Wilson
}
\keyword{datasets}

