% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{remove_language_cols}
\alias{remove_language_cols}
\title{Removing Non-required Language-specific Variables}
\usage{
remove_language_cols(x, lang)
}
\arguments{
\item{x}{data frame to be checked and potentially modified.}

\item{lang}{character, languages to keep. If \code{"all"} no
modifications are done. Else trying to remove
all columns not matching the requested \code{lang}.}
}
\value{
A data frame, potentially modified by removing
all columns not matching the requested language.
}
\description{
Removing Non-required Language-specific Variables
}
\author{
Reto
}
