% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_functions.R
\name{repapply}
\alias{repapply}
\title{Apply Functions Over Replicate Groups}
\usage{
repapply(msdata, ...)
}
\arguments{
\item{msdata}{An object of \code{\link[MSdata]{MSdata}} class.}

\item{fun}{The function to be applied.}
}
\description{
Return a matrix of the same format as intensity matrix of \code{\link[MSdata]{MSdata}}
object with values
}

