% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RTindexing.R
\docType{methods}
\name{RTindexing}
\alias{RTindexing}
\title{Retention time indexing}
\usage{
\S4method{RTindexing}{MSdata}(object, targets.list, mz.window = 0.02,
  rt.window = 30)
}
\arguments{
\item{object}{\code{\link{MSdata-class}} object.
For correct processing there should be "RT" and "MZ" columns in peak data (\code{peakData(object)})}

\item{targets.list}{File path to simple table of the internal standards.
Table should have three columns: compoundID, mass-charge and retention time (with or without header).}

\item{mz.window}{Peak corresponding to standard compound is searched in this range
of MZs around standard MZ value.}

\item{rt.window}{Peak corresponding to standard compound is searched in this range
of RTs around standard RT value.}
}
\value{
\code{\link{MSdata-class}} object with recalculated RT values in peak data
}
\description{
Perform retention time correction in \code{\link{MSdata-class}} object.
Function takes a given list of internal standards, corresponding peaks are found
in data set and basing on their RT shift the shifts for the rest of peaks
are calculated. The resulting data set is realigned.
}
\seealso{
\code{\link{MZindexing}}
}

