% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sg_download_asset}
\alias{sg_download_asset}
\title{Download (Cache) and Import Assets/Data Sets}
\usage{
sg_download_asset(
  x,
  dir = NULL,
  language = c("all", "en", "de", "fr", "it"),
  verbose = FALSE
)
}
\arguments{
\item{x}{data frame (with one row) or list containing
at least the following variables: id, type, href, file_checksum.}

\item{dir}{character, name/path to a directory to cache the
data sets (see 'Details' for more information).}

\item{language}{character, one of \code{"all"} (default), \code{"en"},
\code{"de"}, \code{"fr"}, or \code{"it"}. If set different to \code{"all"}
columns ending in \verb{_<lang>} not matching the requested
language will be removed.}

\item{verbose}{logical, defaults to \code{FALSE}. If set \code{TRUE} the
some messages are printed.}
}
\description{
Download (Cache) and Import Assets/Data Sets
}
\details{
Assets (data sets) provided via geo.admin.ch come
with a file checksum which can be used to check if a file
changed. We use this to cache files.

If a directory is specified (\code{dir}) it will be checked if that file has
already been downloaded and stored in the directory. In this situation, the
local file is used rather than downloading the data. If not, the asset will
be downloaded and stored in the directory specified so that it can be
potentially used again next time.

Please note that the the size of the directory can grow
fast, especially if this mechanic is used when downloading/accessing
data sets that rapidly change (i.e., most recent data) as
a new file will be created every time the checksum changes.
}
\author{
Reto
}
