% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{sin132HRMS}
\alias{sin132HRMS}
\title{Large High Resolution FIE-MS data set}
\usage{
data(sin132HRMS)
}
\description{
\code{sin132HRMS} is a FIE-HRMS data sets contaning 1576 observations across
132 classes.
\itemize{
   \item pos. positive ionisation mode wiith 3079 variables
   \item neg. negative ionisation mode wiith 1452 variables
   \item cls. vector of 132 unique class lables

 }
}
\author{
Tom Wilson
}

