% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{sin132LTQ}
\alias{sin132LTQ}
\title{Large LTQ FIE-MS data set}
\usage{
data(sin132LTQ)
}
\description{
\code{sin132LTQ} is a FIE-MS data sets contaning 1320 observations across
132 classes.
\itemize{
   \item pos. positive ionisation mode wiith 742 variables
   \item neg. negative ionisation mode wiith 688 variables
   \item cls. vector of 132 unique class lables

 }
}
\author{
Tom Wilson
}

