% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{sin132LTQ_raw}
\alias{sin132LTQ_raw}
\title{Large LTQ FIE-MS data set (raw)}
\format{List object containing three elements
	\itemize{
		\item pos positive ionisation mode data frame (1822 variables)
		\item neg negative ionisation mode data frame (1813 variables)
		\item cls vector of class labels (N = 132)
	}}
\usage{
data(sin132LTQ_raw)
}
\description{
Large LTQ FIE-MS data set (raw)
}
\author{
Tom Wilson
}
\keyword{datasets}

