% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Norm_Standards.R
\docType{methods}
\name{StandNorm}
\alias{StandNorm}
\title{Normalisation by standards}
\usage{
\S4method{StandNorm}{MSdata}(msdata, standards.list, mzwindow = 0.01,
  rtwindow = 10, meanInt = 2000, recalculateMean = FALSE)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{standards.list}{The link to the file with the table of standards looking like: compound, m/z, retention time}

\item{mzwindow}{Range (in ppm) for searching the peak corresponding to a standard from the list.}

\item{rtwindow}{Range (in seconds) for searching the peak corresponding to a standard from the list.}

\item{meanInt}{Mean sum intensity of all the standards (usually, got from previous experiments).}

\item{recalculateMean}{If \code{TRUE} then \code{meanInt} is not used, but is recalculated from these particular data.}
}
\value{
\code{\link{MSdata-class}} object with normalised intensity matrix
}
\description{
Normalisation by the list of external or internal standards.
The list of the standards is provided as a file with three columns: compound, m/z, retention time.
By these MZ and RT values corresponding peaks in dataset are determined
and their sum intensity is used for normalisation.\cr
Afterwards, standards are excluded from feature list.
}

