% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EvalMissVal.R
\docType{methods}
\name{EvalMissVal}
\alias{EvalMissVal}
\title{Evaluate Missing Values}
\usage{
\S4method{EvalMissVal}{MSdata}(msdata, method = "Min", ...)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}}}

\item{method}{Method of evaluation, one of:
"Min", "FeatureMin", "FeatureMean", "FeatureMedian", "PCA", "GausSim"}

\item{...}{Other optional parameters:
\itemize{
\item  \code{pcamethods} - character vector containing the name of applied PCA method, one of:

}}
}
\description{
Fill NA positions in data set.
}

