% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessapi.R
\name{ms_api_url}
\alias{ms_api_url}
\title{data.geo.admin.ch API URL}
\usage{
ms_api_url(x = NULL, version = 1L)
}
\arguments{
\item{x}{if \code{NULL} (default) the base URL is returned.
If character, the base URL is extended. If \code{x} is
a vector of length > 1L the elements will be collapsed
using \code{/} to generate a valid URL.}

\item{version}{integer, the API version (defaults to 1L).}
}
\value{
Character of length one with the URL of the API end point.
}
\description{
data.geo.admin.ch API URL
}
\examples{
## Base URL
ms_api_url()
## Base URL with different API version
ms_api_url(version = 2L)

## Extending the base URL to point to a specific API endpoint
ms_api_url("ch.meteoschweiz.ogd-smn")
ms_api_url("/ch.meteoschweiz.ogd-smn")
ms_api_url(c("ch.meteoschweiz.ogd-smn", "items"))

}
\author{
Reto
}
