% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{pw1HRMS}
\alias{pw1HRMS}
\title{Binary FIE-HRMS - (1)}
\usage{
data(pw1HRMS)
}
\description{
\code{pw1HRMS} is a FIE-HRMMS data sets contaning of 24 observations across 2 classes. The two classes are
two distincit accessions from a large scaling profiling experiment.
\itemize{
   \item pos. positive ionisation mode wiith 3079 variables
   \item neg. negative ionisation mode wiith 1422 variables
   \item cls. vector of 2 unique class lables11

 }
}
\author{
Tom Wilson
}

