% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MQI_to_MA_v0.3.R, R/PQI_to_MA_v0.3.R
\name{MQI_to_MA}
\alias{MQI_to_MA}
\alias{PQI_to_MA}
\alias{QI_to_MA}
\title{\code{MQI_to_MA} converts lipid and metabolite QI data output}
\usage{
MQI_to_MA(directory = getwd(), abundance = "Normalised",
  compoundID = "Accepted Compound", facNames = NULL, unite_neg_pos = TRUE)

PQI_to_MA(directory = getwd(), abundance = "Normalised", facNames = NULL,
  compoundID = "shortDescription")
}
\arguments{
\item{directory}{path to the directory containing .csv QI output files.
All the files in all subdirectories are proceeded too.
Remember that in file paths you should use either "/" or "\\", not just "\".}

\item{abundance}{If \code{"Normalised"}, then normalised data are used; if \code{"Raw"}, then raw data are used}

\item{compoundID}{The name of the column in QI output table used as the set of compound IDs.
For \code{MQI_to_MA} one of: \code{"Compound"}, \code{"Accepted Compound ID"}, \code{"Formula"}.\cr
For \code{PQI_to_MA} one of: \code{"Accession"}, \code{"Description"}, \code{"shortDescription"}.\cr
\code{"shortDescription"} means that the data from "Description" column are used,
but they are cut starting with "OS=".}

\item{facNames}{Character vector of grouping factor names.}

\item{unite_neg_pos}{if \code{TRUE}, creates combined file from pos and neg file pairs with matching names;
if \code{FALSE}, just skips this step}
}
\description{
\code{MQI_to_MA} converts lipid and metabolite QI data output

Function for converting QI output .csv files into format convenient for further analysis with MetaboAnalyst.
For proper work, please, make complete QI output files with all the options marked.
Function:
1. Cuts out all the extra data, leaving only columns with abundance values
(either raw or normalised) and a column with compound's IDs.
2. Moves chosen compound ID column to the first position.
3. Expands sample group labels.
4. Separate group labels into grouping factors and creates rows with factors.
5. (for \code{MQI_to_MA}) Merges neg and pos files with matching names.
6. Writes the resulting table into new .csv file.

\code{PQI_to_MA} converts protein QI data output
}
\examples{
PQI_to_MA("//mpimp-golm/user/Homes/Zubkov/QI_data")
PQI_to_MA("\\\\\\\\mpimp-golm\\\\user\\\\Homes\\\\Zubkov\\\\QI_data")
PQI_to_MA("C:/QI_data", abundance = "Raw", compoundID = "Accession")
MQI_to_MA("C:/QI_data", facNames = c("Phenotype", "Treatment", "Time"), unite_neg_pos = FALSE)
}

