% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MQI_to_MA.R, R/PQI_to_MA.R
\name{QI_to_MA}
\alias{MQI_to_MA}
\alias{PQI_to_MA}
\alias{QI_to_MA}
\title{Convert QI to MetaboAnalyst}
\usage{
MQI_to_MA(path = getwd(), abundance = "Raw",
  compoundID = "Accepted Compound", facNames = NULL, unite_neg_pos = TRUE)

PQI_to_MA(path = getwd(), abundance = "Raw", facNames = NULL,
  compoundID = "shortDescription")
}
\arguments{
\item{path}{Character vector. There could be two types of elements:\cr
1) file path(s) to the proceeded .csv QI output file(s).\cr
2) path(s) to the directory(ies) containing .csv files.\cr
In the second case all the files in all subdirectories are proceeded.\cr
Remember that in file paths you should use either "/" or double "\", not just "\".}

\item{abundance}{If \code{"Normalised"}, then normalised data are used; if \code{"Raw"}, then raw data are used}

\item{compoundID}{The name of the column in QI output table used as the set of compound IDs.\cr
For \code{MQI_to_MA} one of: \code{"Compound"}, \code{"Accepted Compound ID"}, \code{"Formula"}.\cr
For \code{PQI_to_MA} one of: \code{"Accession"}, \code{"Description"}, \code{"shortDescription"}.\cr
\code{"shortDescription"} means that the data from "Description" column are used,
but they are cut starting with "OS=".}

\item{facNames}{Character vector of grouping factor names.}

\item{unite_neg_pos}{If \code{TRUE}, creates combined file from "pos" and "neg" file pairs with matching names
if \code{FALSE}, just skips this step}
}
\description{
Function for converting QI output .csv files into format convenient for further analysis with MetaboAnalyst.
For proper work, please, make complete QI output files with all the options marked.\cr
This function:
\enumerate{
  \item Cuts out all the extra data, leaving only columns with abundance values
(either raw or normalised) and a column with compound's IDs.
  \item Moves chosen compound ID column to the first position.
  \item Expands sample group labels.
  \item Separate group labels into grouping factors and creates rows with factors.
  \item (for \code{MQI_to_MA}) Merges neg and pos files with matching names.
  \item Writes the resulting table into new .csv file.
}

\code{MQI_to_MA} converts lipid and metabolite QI data output

\code{PQI_to_MA} converts protein QI data output
}
\examples{
PQI_to_MA("//mpimp-golm/user/Homes/Zubkov/QI_data")
PQI_to_MA("\\\\\\\\mpimp-golm\\\\user\\\\Homes\\\\Zubkov\\\\QI_data")
PQI_to_MA("C:/QI_data", abundance = "Raw", compoundID = "Accession")
MQI_to_MA("C:/QI_data", facNames = c("Phenotype", "Treatment", "Time"), unite_neg_pos = FALSE)
}

