% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{sin132HRMS}
\alias{sin132HRMS}
\title{Large High Resolution FIE-MS data set}
\format{List object containing three elements
	\itemize{
		\item pos postivie ionisation mode data frame (3079 variables)
		\item neg negative ionisation mode data frame (1452 variables)
		\item cls vector of class lables
	}}
\usage{
data(sin132HRMS)
}
\description{
Large High Resolution FIE-MS data set
}
\author{
Tom Wilson
}
\keyword{datasets}

