% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msdata.R
\docType{data}
\name{sin132LTQ}
\alias{sin132LTQ}
\title{Large LTQ FIE-MS data set}
\format{List object containing three elements
	\itemize{
		\item pos postivie ionisation mode data frame (742 variables)
		\item neg negative ionisation mode data frame (688 variables)
		\item cls vector of class lables
	}}
\usage{
data(sin132LTQ)
}
\description{
Large LTQ FIE-MS data set
}
\author{
Tom Wilson
}
\keyword{datasets}

