% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Norm_Standards.R
\docType{methods}
\name{StandNorm}
\alias{StandNorm}
\title{Normalisation by standards}
\usage{
\S4method{StandNorm}{MSdata}(msdata, standards.list, meanInt = 2000,
  recalculateMean = FALSE)
}
\arguments{
\item{msdata}{\code{\link{MSdata-class}} object}

\item{standards.list}{The link to the file with the table of standards looking like: compound, m/z, retention time}

\item{meanInt}{Mean sum intensity of all the standards (usually, got from previous experiments).}

\item{recalculateMean}{If \code{TRUE} then \code{meanInt} is not used, but is recalculated from these particular data.}
}
\value{
\code{\link{MSdata-class}} object with normalised intensity matrix
}
\description{
Normalisation by the list of external or internal standards.
In each sample all the intensities are normalized by sum of standards intensities.
Afterwards, standards are excluded from feature list.
}

